Add files via upload
A marketing company wants to estimate the amount of spending of its customers by analyzing customer spending behavior. The company wants to estimate the amount of spending by customers using factors such as the population of the city where the customers live, the amount of past spending, their reaction to internet advertisements and the time they spend on the website.
Based on data, the marketing company wants to estimate the amount of spending by customers by performing multiple linear regression analysis.